# product-sale-prediction-AI/resources/openapi.yaml
openapi: 3.0.3
info:
  title: E-commerce Product Prediction API
  description: |
    Intelligent E-commerce API with vector search, sales analytics, and prediction capabilities.
    This API provides advanced features for product management, search, analytics, and sales predictions.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development Server
  - url: https://api.example.com
    description: Production Server
...
components:
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: api-key
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          example: "P123456"
          description: Unique product identifier
        name:
          type: string
          example: "Wireless Bluetooth Headphones"
          description: Product name
        category:
          type: string
          example: "Electronics"
          description: Product category
        brand:
          type: string
          example: "AudioTech"
          description: Product brand
        price:
          type: number
          format: double
          example: 99.99
          description: Product price
        createTimestamp:
          type: string
          format: date-time
          example: "2023-05-15T10:30:00"
          description: Product creation timestamp
        description:
          type: string
          example: "High-quality wireless headphones with noise cancellation"
          description: Detailed product description
...
    SalesSearchRequest:
      type: object
      properties:
        topN:
          type: integer
          example: 10
          description: Number of top products to retrieve
        startTime:
          type: string
          example: "2023/01"
          description: Start time in year/month format
        endTime:
          type: string
          example: "2023/12"
          description: End time in year/month format
        sellerId:
          type: string
          example: "SELLER789"
          description: Seller ID to filter by
        category:
          type: string
          example: "Electronics"
          description: Category to filter by
      required:
        - topN
        - startTime
        - sellerId
    SalesPredictionRequest:
      type: object
      properties:
        productId:
          type: string
          example: "P123456"
          description: Product ID to predict sales for
        sellerId:
          type: string
          example: "SELLER789"
          description: Seller ID
        startTime:
          type: string
          example: "2023/01"
          description: Start time in year/month format
        endTime:
          type: string
          example: "2023/12"
          description: End time in year/month format
      required:
        - productId
        - sellerId
        - startTime
paths:
  /v1/products:
    get:
      tags:
        - Products
      summary: Get all products
      description: Retrieve all products, optionally filtered by category and/or seller ID
      operationId: getAllProducts
      parameters:
        - name: category
          in: query
          description: Filter by product category
          required: false
          schema:
            type: string
          example: "Electronics"
        - name: sellerId
          in: query
          description: Filter by seller ID
          required: false
          schema:
            type: string
          example: "SELLER789"
      responses:
        '200':
          description: Successfully retrieved products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  count:
                    type: integer
                    example: 25
                  category:
                    type: string
                    example: "Electronics"
                  sellerId:
                    type: string
                    example: "SELLER789"
  /v1/product/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Retrieve a product by its unique identifier
      operationId: getProductById
      parameters:
        - name: id
          in: path
          description: Product ID
          required: true
          schema:
            type: string
          example: "P123456"
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                type: object
                properties:
                  product:
                    $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
...
  /v1/sales/search:
    post:
      tags:
        - Analytics
      summary: Search for top selling products
      description: Find top selling products by time range, category, and seller ID
      operationId: searchTopSellingProducts
      requestBody:
        description: Search parameters
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalesSearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      type: object
                      properties:
                        productId:
                          type: string
                          example: "P123456"
                        totalQuantity:
                          type: integer
                          example: 250
                        name:
                          type: string
                          example: "Wireless Bluetooth Headphones"
                        category:
                          type: string
                          example: "Electronics"
                        brand:
                          type: string
                          example: "AudioTech"
                        price:
                          type: number
                          format: double
                          example: 99.99
                  count:
                    type: integer
                    example: 10
                  query:
                    $ref: '#/components/schemas/SalesSearchRequest'
        '400':
          description: Bad request due to invalid parameters
  /v1/sales/predict:
    post:
      tags:
        - Predictions
      summary: Predict product sales
      description: Predict future sales for a specific product
      operationId: predictProductSales
      requestBody:
        description: Prediction parameters
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalesPredictionRequest'
      responses:
        '200':
          description: Sales prediction
          content:
            application/json:
              schema:
                type: object
                properties:
                  prediction:
                    type: object
                    additionalProperties:
                      type: object
                    example:
                      "2023/06": 
                        quantity: 45
                        revenue: 4495.5
                      "2023/07": 
                        quantity: 52
                        revenue: 5199.48
                  product:
                    $ref: '#/components/schemas/Product'
                  query:
                    $ref: '#/components/schemas/SalesPredictionRequest'
        '400':
          description: Bad request due to invalid parameters
        '404':
          description: Product not found
...